+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+- check-dns.rb: Added ability to use many name servers (@yuri-zubov)
 
 ## [1.3.0] - 2017-11-04
 ### Added